requirements.txt
asgiref==3.5.2
bcrypt==3.2.2
black==22.3.0
cffi==1.15.0
click==8.1.3
Django==4.0.5
djangorestframework==3.13.1
mypy-extensions==0.4.3
mysqlclient==2.1.0
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.2
pycparser==2.21
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.2.0
